resolve RQ01-Montos Decimales + pruebas

+++ Cuenta.cs
@@ -20,7 +20,7 @@ namespace ATMSim
         public TipoCuenta Tipo { get; private set; }
         public string Numero { get; private set; }
         int montoSobregiro;
-        int monto = 0;
+        double monto = 0;
         public int MontoSobregiro{
             get {
                 if(this.Tipo == TipoCuenta.Corriente){
@@ -38,7 +38,7 @@ namespace ATMSim
                 }
             }
         }
-        public int Monto { 
+        public double Monto { 
             get { return monto; } 
             set 
             {
@@ -49,7 +49,7 @@ namespace ATMSim
             } 
         }
 
-        public Cuenta(string numero, TipoCuenta tipo, int monto = 0, int montoSobregiro = 0) 
+        public Cuenta(string numero, TipoCuenta tipo, double monto = 0, int montoSobregiro = 0) 
         {
             if (!Regex.Match(numero, @"[0-9]+").Success)
                 throw new ArgumentException("Numero de cuenta inválido");
